# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGI Judd's Idea Catalog API",
    "description": "Transform LLM conversations into actionable Linear projects",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/conversations": {
      "get": {
        "summary": "List Conversations",
        "description": "Get a paginated list of all imported conversations.",
        "operationId": "list_conversations_api_conversations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConversationListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response List Conversations Api Conversations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{conversation_id}": {
      "get": {
        "summary": "Get Conversation",
        "description": "Get a specific conversation by ID with complete details.",
        "operationId": "get_conversation_api_conversations__conversation_id__get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConversationResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Get Conversation Api Conversations  Conversation Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Conversation",
        "description": "Delete a specific conversation by ID.",
        "operationId": "delete_conversation_api_conversations__conversation_id__delete",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/MessageResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Delete Conversation Api Conversations  Conversation Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Conversation",
        "description": "Update a conversation's title.",
        "operationId": "update_conversation_api_conversations__conversation_id__patch",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConversationUpdateResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Update Conversation Api Conversations  Conversation Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations/{conversation_id}/files": {
      "post": {
        "summary": "Upload File",
        "description": "Upload a file attachment for a conversation.\n\nArgs:\n    conversation_id: ID of the conversation\n    file: File to upload (multipart/form-data)\n    llm_model: LLM model to use for processing\n    llm_provider: LLM provider to use for processing\n\nReturns:\n    JSON response with file metadata\n\nRaises:\n    HTTPException: If validation fails or upload fails",
        "operationId": "upload_file_api_conversations__conversation_id__files_post",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_conversations__conversation_id__files_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/FileUploadResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Upload File Api Conversations  Conversation Id  Files Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Conversation Files",
        "description": "List all file attachments for a conversation.\n\nArgs:\n    conversation_id: ID of the conversation\n\nReturns:\n    JSON response with list of file attachments\n\nRaises:\n    HTTPException: If conversation not found",
        "operationId": "list_conversation_files_api_conversations__conversation_id__files_get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/FileListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response List Conversation Files Api Conversations  Conversation Id  Files Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/files/{file_id}/download": {
      "get": {
        "summary": "Download File",
        "description": "Download a file attachment via temporary signed URL.\n\nArgs:\n    file_id: ID of the file attachment\n\nReturns:\n    Redirect response to S3 signed URL\n\nRaises:\n    HTTPException: If file not found or download fails",
        "operationId": "download_file_api_conversations_files__file_id__download_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_api_conversations__conversation_id__files_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model"
          },
          "llm_provider": {
            "type": "string",
            "title": "Llm Provider"
          }
        },
        "type": "object",
        "required": [
          "file",
          "llm_model",
          "llm_provider"
        ],
        "title": "Body_upload_file_api_conversations__conversation_id__files_post"
      },
...
      "ConversationListResponse": {
        "properties": {
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationListItem"
            },
            "type": "array",
            "title": "Conversations",
            "description": "List of conversations"
          }
        },
        "type": "object",
        "required": [
          "conversations"
        ],
        "title": "ConversationListResponse",
        "description": "Response for listing conversations."
      },
      "ConversationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Database ID of the conversation"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Original conversation share URL"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Conversation title"
          },
          "import_date": {
            "type": "string",
            "title": "Import Date",
            "description": "ISO format import timestamp"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO format creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO format last update timestamp"
          },
          "is_locked": {
            "type": "boolean",
            "title": "Is Locked",
            "description": "Whether the conversation is locked"
          },
          "has_images": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Images",
            "description": "Whether conversation contains images"
          },
          "has_pdfs": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Pdfs",
            "description": "Whether conversation contains PDFs"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "ID of the user who imported the conversation"
          },
          "user_name": {
            "type": "string",
            "title": "User Name",
            "description": "Name of the user who imported the conversation"
          },
          "user_email": {
            "type": "string",
            "title": "User Email",
            "description": "Email of the user who imported the conversation"
          },
          "imported_chat": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ImportedChatMessage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imported Chat",
            "description": "Conversation messages (optional)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "title",
          "import_date",
          "created_at",
          "updated_at",
          "is_locked",
          "user_id",
          "user_name",
          "user_email"
        ],
        "title": "ConversationResponse",
        "description": "Response model for conversation API endpoints."
      },
      "ConversationUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "New title for the conversation"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ConversationUpdate",
        "description": "Data required to update a conversation."
      },
      "ConversationUpdateResponse": {
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ConversationResponse",
            "description": "Updated conversation"
          }
        },
        "type": "object",
        "required": [
          "conversation"
        ],
        "title": "ConversationUpdateResponse",
        "description": "Response for conversation updates."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response schema."
      },
...
      "FileListResponse": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileAttachment"
            },
            "type": "array",
            "title": "Files",
            "description": "List of file attachments"
          },
          "file_count": {
            "type": "integer",
            "title": "File Count",
            "description": "Number of files"
          }
        },
        "type": "object",
        "required": [
          "files",
          "file_count"
        ],
        "title": "FileListResponse",
        "description": "File list response."
      },
      "FileMetadata": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "File attachment ID"
          },
          "s3_key": {
            "type": "string",
            "title": "S3 Key",
            "description": "S3 storage key"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Original filename"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size",
            "description": "File size in bytes"
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "description": "MIME type"
          },
          "conversation_id": {
            "type": "integer",
            "title": "Conversation Id",
            "description": "Associated conversation ID"
          }
        },
        "type": "object",
        "required": [
          "id",
          "s3_key",
          "filename",
          "file_size",
          "file_type",
          "conversation_id"
        ],
        "title": "FileMetadata",
        "description": "File metadata response."
      },
      "FileUploadResponse": {
        "properties": {
          "file": {
            "$ref": "#/components/schemas/FileMetadata",
            "description": "Uploaded file metadata"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Success message"
          }
        },
        "type": "object",
        "required": [
          "file",
          "message"
        ],
        "title": "FileUploadResponse",
        "description": "File upload response."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportedChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Message role: 'user' or 'assistant'"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Message content as Markdown"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ImportedChatMessage",
        "description": "Represents a single message in a conversation."
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Simple message response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}